feat: Basic WebRTC types, improved build scripts
* added a gitignore to ignore node modules

* clean up build process

This build directory made no sense to me. It was requiring a non
existent file and writing to an un-used/ un-exported file.  Also the
source file was not exported or used anywhere that I could see. To make
building better I made the following changes:

- rename the `build` directory to `scripts` as `build` is misleading
- remove the file `source.json`
- propagate the last pull request #4 to the build script
- add an npm command to build `npm run build`

Now running `npm run build` will use the file `globals-docs.json` as its
input and output file.  To add more browser features just set a key to
the feature name and the value to https://developer.mozilla.org and run
`npm run build` to resolve the mozilla documentation and generate a new
`globals-docs.json`

* add support for the basic webrtc types

* refactor build script to resolve issue #5

* remove wildcard .log ignore

* remove `SVGColor` global as no browser supports it